+++ lca_activity_browser/app/ui/network/sankey.py
@@ -2,17 +2,17 @@
 import os
 import json
 import collections
-from concurrent.futures import ThreadPoolExecutor
 from PyQt5 import QtWidgets, QtCore, QtWebEngineWidgets, QtWebChannel
 import numpy as np
 import brightway2 as bw
 import matplotlib.pyplot as plt
 from .signals import sankeysignals
+from .worker_threads import gt_worker_thread
+
 
 class SankeyWindow(QtWidgets.QMainWindow):
     def __init__(self, parent=None):
         super().__init__(parent)
-
         self.sankeywidget = SankeyWidget(self)
         self.setCentralWidget(self.sankeywidget)
         self.show()
@@ -30,7 +30,8 @@ class SankeyWidget(QtWidgets.QWidget):
         self.grid_lay.addWidget(QtWidgets.QLabel('Method: '), 1, 0)
         self.cs = self.parent().parent().list_widget.name
         self.func_units = bw.calculation_setups[self.cs]['inv']
-        self.func_units = [{bw.get_activity(k):v for k,v in fu.items()} for fu in self.func_units]
+        self.func_units = [{bw.get_activity(k):v for k,v in fu.items()} 
+                           for fu in self.func_units]
         self.methods = bw.calculation_setups[self.cs]['ia']
         self.func_unit_cb = QtWidgets.QComboBox()
         self.func_unit_cb.addItems(
@@ -88,7 +89,7 @@ class SankeyWidget(QtWidgets.QWidget):
 
         self.func_unit_cb.currentIndexChanged.connect(self.new_sankey)
         self.method_cb.currentIndexChanged.connect(self.new_sankey)
-        self.color_attr_cb.currentIndexChanged.connect(self.new_sankey)
+        self.color_attr_cb.currentIndexChanged.connect(self.update_colors)
         self.cutoff_sb.valueChanged.connect(self.new_sankey)
         
         # connections
@@ -102,6 +103,12 @@ class SankeyWidget(QtWidgets.QWidget):
         color_attr = self.color_attr_cb.currentText()
         cutoff = self.cutoff_sb.value()
         self.sankey = SankeyGraphTraversal(demand, method, cutoff, color_attr)
+
+    def update_colors(self):
+        self.sankey.color_attr = self.color_attr_cb.currentText()
+        self.sankey.colors()
+        self.sankey.to_json()
+        self.bridge.lca_calc_finished.emit(self.sankey.json_data)
 
     def draw_sankey(self):
         self.view.load(self.url)
@@ -132,21 +139,6 @@ class Bridge(QtCore.QObject):
     def viewer_ready(self):
         self.viewer_waiting.emit()
 
-       
-class GraphTraversalThread(QtCore.QThread):
-    def update_params(self, demand, method, cutoff, max_calc):
-        self.demand = demand
-        self.method = method
-        self.cutoff = cutoff
-        self.max_calc = max_calc
-        
-    def run(self):
-        res = bw.GraphTraversal().calculate(self.demand, self.method, self.cutoff, self.max_calc)
-        sankeysignals.gt_ready.emit(res)
-
-
-gt_worker_thread = GraphTraversalThread()
-
 
 class SankeyGraphTraversal:
     def __init__(self, demand, method, cutoff=0.005, color_attr='flow'):
@@ -158,7 +150,8 @@ class SankeyGraphTraversal:
         
     def init_graph(self, gt):
         self.nodes = []
-        self.reverse_activity_dict = {v:k for k,v in gt['lca'].activity_dict.items()}
+        self.reverse_activity_dict = {v:k for k,v in 
+                                      gt['lca'].activity_dict.items()}
         self.edges = gt['edges']
         self.root_score = gt['nodes'][-1]['cum']
         self.expanded_nodes = set()
@@ -167,14 +160,14 @@ class SankeyGraphTraversal:
             self.expand(self.links[0]['target'])
         sankeysignals.initial_sankey_ready.emit()
 
-
     def expand(self, ind):
         if ind in self.expanded_nodes:
             self.expanded_nodes.remove(ind)
             self.remove_dangling_nodes()
         else:
             self.expanded_nodes.add(ind)
-        displayed_edges = [e for e in self.edges if e['to'] in self.expanded_nodes]
+        displayed_edges = [e for e in self.edges if 
+                           e['to'] in self.expanded_nodes]
         self.links = [{'source': e['to'], 
                        'target': e['from'], 
                        'value': e['impact'],
@@ -183,12 +176,16 @@ class SankeyGraphTraversal:
             {li['target'] for li in self.links})
         self.nodes = [{'id':n, 'style': 'process'} for n in self.nodes_set]
         self.colors()
+        self.to_json()
+
+    def to_json(self):
         sankey_dict = {'links':self.links, 'nodes':self.nodes}
         self.json_data = json.dumps(sankey_dict)
 
     def remove_dangling_nodes(self):
         while True:
-            displayed_edges = [e for e in self.edges if e['to'] in self.expanded_nodes]
+            displayed_edges = [e for e in self.edges if 
+                               e['to'] in self.expanded_nodes]
             to = {e['to'] for e in displayed_edges}
             from_ = {e['from'] for e in displayed_edges}
             dangling = to.difference(from_)
@@ -222,12 +219,15 @@ class SankeyGraphTraversal:
         options = sorted(
             {self.get_bw_activity_by_index(n).get(self.color_attr) for 
              n in self.nodes_set.difference({-1})})
-        color_dict = {o:self.viridis_r_hex(v) for o,v in zip(options, np.linspace(0,1,len(options)))}
+        color_dict = {o:self.viridis_r_hex(v) for o,v in 
+                      zip(options, np.linspace(0,1,len(options)))}
         for link in self.links:
-            link['color'] = color_dict[self.get_bw_activity_by_index(link['target']).get(self.color_attr)]
+            link['color'] = color_dict[self.get_bw_activity_by_index(
+                link['target']).get(self.color_attr)]
 
     @staticmethod
     def viridis_r_hex(v):
-        return "#{0:02x}{1:02x}{2:02x}".format(*plt.cm.viridis_r(v, bytes=True)[:-1])
+        return "#{0:02x}{1:02x}{2:02x}".format(
+            *plt.cm.viridis_r(v, bytes=True)[:-1])
